\name{rth-package}
\alias{rth-package}
\docType{package}
\title{rth}
\description{
  TODO
}
\details{
\tabular{ll}{
  Package: \tab rth\cr
  Type: \tab Package\cr
  License: \tab GPL >= 2\cr
}
}
\references{
  \url{http://heather.cs.ucdavis.edu/~matloff/rth.html}
}
\author{
  Norm Matloff
}
\keyword{Package}
