% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{extract_rth_md}
\alias{extract_rth_md}
\title{Extract Tick History Market Depth on demand}
\usage{
extract_rth_md(
  identifier = "0#TFMB:",
  identifier_type = "ChainRIC",
  timestamp_in = "GmtUtc",
  query_start_datetime = Sys.time() - 72 * 60 * 60,
  query_end_datetime = Sys.time() - 48 * 60 * 60,
  md_view = "NormalizedLL2",
  number_of_levels = 10,
  display_source_ric = TRUE,
  result_path = fs::path_home("Downloads", sprintf(fmt = "RTH_MD_extractum_\%i",
    as.integer(Sys.time())), ext = "csv.gz"),
  overwrite = FALSE
)
}
\arguments{
\item{identifier}{the required RIC or ChainRIC value(s)}

\item{identifier_type}{the corresponding type of the identifier(s) ("Ric" or "ChainRIC")}

\item{timestamp_in}{timestamps will be displayed in "GmtUtc" or
"LocalExchangeTime"}

\item{query_start_datetime}{the range's start date and time in POSIXct format}

\item{query_end_datetime}{the range's end date and time in POSIXct format}

\item{md_view}{market depth view, value can be:
"LegacyLevel2", "NormalizedLL2", "RawMarketByOrder", "RawMarketByPrice" or
"RawMarketMaker"; defaults to "NormalizedLL2"}

\item{number_of_levels}{in case of "NormalizedLL2" view, defaults to 10}

\item{display_source_ric}{if TRUE, the result table will have an extra
field which indicates the underlying instrument}

\item{result_path}{path of the prospective result file,
the extension must be ".csv.gz"}

\item{overwrite}{should overwrite existing result file TRUE/FALSE}
}
\value{
The result format is the native/raw result from the
underlying extractor (usually csv).
}
\description{
If the response is available instantly, it performs an
on demand extraction which returns the raw results.
Otherwise the server accepts the request and provides "monitor URL".
In the latter case the extraction status/end result can be polled
with poll_async_status() function using the "monitor URL" as
an argument.
}
