% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{extract_rth_tas}
\alias{extract_rth_tas}
\title{Extract Tick History Time And Sales on demand}
\usage{
extract_rth_tas(
  content_field_names = grep(pattern = "^Trade.+(Price|Size)", x =
    get_valid_content_field_names(report_template_type = "TickHistoryTimeAndSales")$Name,
    value = TRUE),
  identifier = "0#TFMB:",
  identifier_type = "ChainRIC",
  appl_corr_and_canc = TRUE,
  timestamp_in = "GmtUtc",
  query_start_datetime = Sys.time() - 72 * 60 * 60,
  query_end_datetime = Sys.time() - 48 * 60 * 60,
  display_source_ric = TRUE,
  result_path = fs::path_home("Downloads", sprintf(fmt = "RTH_TAS_extractum_\%i",
    as.integer(Sys.time())), ext = "csv.gz"),
  overwrite = FALSE
)
}
\arguments{
\item{content_field_names}{set the field names of the result table}

\item{identifier}{the required RIC or ChainRIC value(s)}

\item{identifier_type}{the corresponding type of the identifier(s)}

\item{appl_corr_and_canc}{if TRUE, the result table will display the
original and the corrected values as well}

\item{timestamp_in}{timestamps will be displayed in "GmtUtc" or
"LocalExchangeTime"}

\item{query_start_datetime}{the range's start date and time in POSIXct format}

\item{query_end_datetime}{the range's end date and time in POSIXct format}

\item{display_source_ric}{if TRUE, the result table will have an extra
field which indicates the underlying instrument}

\item{result_path}{path of the prospective result file,
the extension must be ".csv.gz"}

\item{overwrite}{should overwrite existing result file TRUE/FALSE}
}
\value{
The result format is the native/raw result from the
underlying extractor (usually csv).
}
\description{
If the response is available instantly, it performs an
on demand extraction which returns the raw results.
Otherwise the server accepts the request and provides "monitor URL".
In the latter case the extraction status/end result can be polled
with poll_async_status() function using the "monitor URL" as
an argument.
}
