% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_active_jobs}
\alias{get_active_jobs}
\title{Retrieve the active jobs}
\usage{
get_active_jobs()
}
\value{
Returns the in flight jobs for all asynchronous requests.
}
\description{
Retrieve the active jobs
}
