% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_all_subscriptions}
\alias{get_all_subscriptions}
\title{Retrieves all Subscriptions of the Authorized User.}
\usage{
get_all_subscriptions()
}
\value{
Returns Collection of all Subscriptions
}
\description{
Retrieves all Subscriptions of the Authorized User.
}
