% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_completed_jobs}
\alias{get_completed_jobs}
\title{Retrieve the completed jobs}
\usage{
get_completed_jobs()
}
\value{
Returns the ready jobs for all asynchronous requests.
}
\description{
Retrieve the completed jobs
}
