% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_extr_file_by_job_id}
\alias{get_extr_file_by_job_id}
\title{Retrieve the extracted file details by job_id}
\usage{
get_extr_file_by_job_id(job_id)
}
\arguments{
\item{job_id}{the job ID of the extracted file}
}
\value{
Returns the details of the extracted file.
}
\description{
Retrieve the extracted file details by job_id
}
