% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_instr_list_details}
\alias{get_instr_list_details}
\title{Retrieves the instrument list matching the specified name}
\usage{
get_instr_list_details(instr_list_name)
}
\arguments{
\item{instr_list_name}{name of the instrument list}
}
\value{
the instrument with list id, name and timestamps
Note that the search for the specified list name is case sensitive.
}
\description{
Retrieves the instrument list matching the specified name
}
