% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_instr_lists}
\alias{get_instr_lists}
\title{Retrieves the instrument lists of the authenticated user}
\usage{
get_instr_lists()
}
\value{
the instrument lists with id, name and timestamps
Note that the search for the specified list name is case sensitive.
}
\description{
Retrieves the instrument lists of the authenticated user
}
