% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_package_list}
\alias{get_package_list}
\title{Retrieve the list of all packages}
\usage{
get_package_list()
}
\value{
Return the list of package id, subscription id,
package name and description
}
\description{
Retrieve the list of all packages
}
