% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_quota_info}
\alias{get_quota_info}
\title{Retrieve the user's quota information}
\usage{
get_quota_info()
}
\value{
Returns the quota summary for the logged in user.
}
\description{
Quota limitations apply at the client level.
All the users pertaining to a client have the same quota restrictions.
}
