% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_ric_list}
\alias{get_ric_list}
\title{Retrieve the authorized and extracted RIC list}
\usage{
get_ric_list(category_code = "Futures")
}
\value{
Returns the RIC list that the user is authorized to use.
}
\description{
If the quota service sees a new RIC,
it gets added to the authorized list,
unless the cummulative quota has been reached for that asset class
}
