% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_token}
\alias{get_token}
\title{Request and cache authentication token}
\usage{
get_token(
  uname = Sys.getenv("rth_dss_username"),
  pword = Sys.getenv("rth_dss_password")
)
}
\arguments{
\item{uname}{Refinitiv Tick History DataScope Select username}

\item{pword}{Refinitiv Tick History DataScope Select password}
}
\value{
An authentication token invisibly and cache it for further usage.
}
\description{
Request and cache authentication token
}
