% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_user_info}
\alias{get_user_info}
\title{Retrieves user information}
\usage{
get_user_info(uname = Sys.getenv("rth_dss_username"))
}
\arguments{
\item{uname}{Refinitiv Tick History DataScope Select username}
}
\value{
Return list of UserId, UserName, Email and Phone
}
\description{
Retrieves user information
}
