% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{get_valid_content_field_names}
\alias{get_valid_content_field_names}
\title{Retrieve the list of valid ContentFieldTypes for the report template type.}
\usage{
get_valid_content_field_names(report_template_type = "TickHistoryTimeAndSales")
}
\arguments{
\item{report_template_type}{Available template type for RTH are
"TickHistoryTimeAndSales","TickHistoryMarketDepth",
and "TickHistoryIntradaySummaries".}
}
\value{
Returns a list of valid content field names for
the report template type.
}
\description{
Retrieve the list of valid ContentFieldTypes for the report template type.
}
