% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{historical_search}
\alias{historical_search}
\title{Search for historical instruments given an instrument identifier.}
\usage{
historical_search(
  identifier = "0#N1BM:",
  identifier_type = "ChainRIC",
  start_datetime = Sys.time() - 72 * 60 * 60,
  end_datetime = Sys.time(),
  results_by = "Ric"
)
}
\arguments{
\item{identifier}{Instrument identifier}

\item{identifier_type}{The type of identifier.
Supported types are ChainRIC, RIC, ISIN, CUSIP, SEDOL.
Search will look for the identifier in all supported
types when not specified.}

\item{start_datetime}{The range's start date and time in POSIXct format.}

\item{end_datetime}{The range's end date and time in POSIXct format.}

\item{results_by}{Determines what information is returned
for each found RIC.
"Ric": Returns information purely based on the RIC history.
"Entity": Returns information based on the entity associated
with the "RIC" on the end date of the range.
This will cause RIC rename history to be returned.
Defaults to "Ric".}
}
\value{
Return instruments may be currently active, or inactive
'historical only' instruments.
}
\description{
Search for historical instruments given an instrument identifier.
}
