% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{poll_async_status}
\alias{poll_async_status}
\title{Poll the asynchronous extraction status.}
\usage{
poll_async_status(
  monitor_url,
  result_path = fs::path_home("Downloads", sprintf(fmt = "RTH_extractum_\%i",
    as.integer(Sys.time())), ext = "csv.gz"),
  overwrite = FALSE
)
}
\arguments{
\item{monitor_url}{the monitor URL}

\item{result_path}{path of the prospective result file,
the extension must be ".csv.gz"}

\item{overwrite}{should overwrite existing result file TRUE/FALSE}
}
\value{
The "monitor_url" status message and
the result file downloaded, if applicable.
}
\description{
On Demand extraction requests are executed
asynchronously, which means as soon as possible.
This command checks the readiness of the job
and downloads the extractum if that is ready.
}
