% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTH.R
\name{reference_history}
\alias{reference_history}
\title{Retrieve FID reference history events for a set of RICs
in a specified date range.}
\usage{
reference_history(
  ric_list = c("0#N1BM:", "0#NGMM:"),
  start_datetime = Sys.time() - 72 * 60 * 60,
  end_datetime = Sys.time()
)
}
\arguments{
\item{ric_list}{the RIC identifiers to return reference history for}

\item{start_datetime}{the range's start date and time in POSIXct format}

\item{end_datetime}{the range's end date and time in POSIXct format}
}
\value{
list of ReferenceHistoryResult tables
}
\description{
Retrieve FID reference history events for a set of RICs
in a specified date range.
}
