% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthdist.R
\name{rthdist}
\alias{rthdist}
\alias{rthdis}
\title{Parallel Computation of Distances}
\usage{
rthdist(m, nthreads = rth.nthreads())
}
\arguments{
\item{m}{Matrix/data frame for between-rows distance computation.}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\value{
Object of R class \code{"matrix"}.  Duplicate computation is avoided
in \code{rthdist()} but the entire matrix is constructed.
}
\description{
Similar to R's \code{dist()}.
}
\details{
Calling \code{rthdist()} on the matrix \code{m} returns the euclidean
distances between all pairs of rows in \code{m}.
}
\examples{

\dontrun{
library(Rth)
m1 <- rbind(1:2,3:4)
m1
rthdist(m1)
}

}
