% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rthma}
\alias{rthma}
\title{Parallel Moving Average}
\usage{
rthma(x, w, nthreads = rth.nthreads())
}
\arguments{
\item{x}{a numeric vector}

\item{w}{window width for the moving average}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\value{
The moving-average vector.  Element i is the average of
\code{x[i-w+1]},..., \code{x[i]}.
}
\description{
Computes a moving-average vector
}
\details{
Computes a vector like that of \code{runmean()} in \pkg{caTools}, except
that only full windows are computed, causing a discrepancy at the endpoints
with \code{runmean()}.
}
\examples{
\dontrun{
x <- c(5,2,12,13,8,0,50)
rthma(x,3)
# should output c(6.333333,9.000000,11.000000,7.000000,19.333333)
}
}

