% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthmean.R
\name{rthmean}
\alias{rthmean}
\title{Calculate the mean of a one dimensional vector}
\usage{
rthmean(x, nthreads = rth.nthreads())
}
\arguments{
\item{x}{The data array}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\value{
The mean of the array
}
\description{
Computes the mean like \code{mean()}
}
\examples{

\dontrun{
x <- c(1:10, 50)
x
rthmean(x)
}

}
\author{
Norm Matloff <matloff@cs.ucdavis.edu> and Drew Schmidt
<schmidt@math.utk.edu>
}
