% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthnorm.R
\name{rthnorm}
\alias{rthnorm}
\title{Parallel Vector Norm}
\usage{
rthnorm(x, p = 2, nthreads = rth.nthreads())
}
\arguments{
\item{x}{a numeric vector}

\item{p}{type of norm to compute}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\value{
Norm of a vector
}
\description{
Compuetes the norm of a vector
}
\details{
Computes the norm of a vector similar to \code{norm()}
}
\examples{

\dontrun{
x <- c(5,2,12,13,8,0,50)
x
rthnorm(x)
}

}
