% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthpearson.R
\name{rthpearson}
\alias{rthpearson}
\title{Parallel Pearson Product Moment Correlation}
\usage{
rthpearson(x, y, nthreads = rth.nthreads())
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\value{
The Pearson Moment Correlation
}
\description{
Computes Pearson Product Moment Correlation
}
\examples{

\dontrun{
 x <- c(44.4, 45.9, 41.9, 53.3, 44.7, 44.1, 50.7, 45.2, 60.1)
 y <- c( 2.6,  3.1,  2.5,  5.0,  3.6,  4.0,  5.2,  2.8,  3.8)
 rthpearson(x, y)
}
}
