% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rthrnorm}
\alias{rthrnorm}
\title{Parallel Random Normal Generation}
\usage{
rthrnorm(n, mean = 0, sd = 1, seed = rth.getseed(),
  nthreads = rth.nthreads())
}
\arguments{
\item{n}{The number of values to generate}

\item{mean,sd}{Parameters for normal random variables.}

\item{seed}{Seed for the random number generation.}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\description{
Generate from a normal distribution in parallel.
}
\references{
\url{https://thrust.github.io/doc/group__predefined__random.html}
}

