% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthxpos.R
\name{rthxpos}
\alias{rthxpos}
\title{Parallel Matrix Transpose}
\usage{
rthxpos(m)
}
\arguments{
\item{m}{a matrix to transpose}
}
\value{
the transposed matrix
}
\description{
Transposes a matrix
}
\details{
Computes the transpose of a vector similar to  \code{t()}
}
\examples{

\dontrun{
m <- matrix(runif(30), 5, 6)
rthxpos(m)
}

}
