\name{rth-package}
\alias{rth-package}
\docType{package}
\title{
  Flexible parallel computing from R, on multicore machines and GPUs.
}
\description{
  A collection of R wrappers to C++ functions written in the Thrust
  environment.  Since the latter can be configured for execution either on
  multicore systems or on graphics processing units (GPUs), the same code
  works on both types of platforms.
  
  R programmers may use these functions without knowing how to program in
  C++, and without knowledge of the underlying parallel hardware.
}
\details{
\tabular{ll}{
  Package: \tab rth\cr
  Type: \tab Package\cr
  License: \tab GPL >= 2\cr
}
}
\references{
  \url{http://heather.cs.ucdavis.edu/~matloff/rth.html}
}
\author{
  Norm Matloff
}
\keyword{Package}
