% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rthcolsums}
\alias{rthcolsums}
\title{Matrix column sums.}
\usage{
rthcolsums(m, nthreads = rth.nthreads())
}
\arguments{
\item{m}{The data, in data frame or matrix form.}

\item{nthreads}{An optional argument for specifying the number of threads
(in the case of OpenMP or TBB backends) that the routine should (at most)
use.  Does nothing with a CUDA backend.  See \link{nthreads}.}
}
\value{
Vector of column sums.
}
\description{
For the given input matrix, returns the vector of column sums.
}
\examples{
\dontrun{

# "attitude" data set, built-in to R
rthcolsums(attitude)
# should print 1939 1998...
}
}
\author{
Norm Matloff <matloff@cs.ucdavis.edu> and Drew Schmidt
<schmidt@math.utk.edu>
}

